requirements.txt
asgiref==3.5.1
certifi==2022.5.18.1      
charset-normalizer==2.0.12
colorama==0.4.5
commonmark==0.9.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0
django-cors-headers==3.12.0
django-crispy-forms==1.7.2
django-extensions==3.1.5
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.0
httpie==3.2.1
idna==3.3
importlib-metadata==4.11.4
is-safe-url==1.0
Markdown==3.3.7
multidict==6.0.2
numpy==1.22.3
Pillow==9.1.0
platformdirs==2.5.2
Pygments==2.12.0
PyJWT==2.4.0
PySocks==1.7.1
pytz==2018.7
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
zipp==3.8.0